refactor ♻️ menu modal option 부분 컴포넌트 분리
